Inicio del CRUD

+++ routes/web.php
@@ -27,13 +27,14 @@ Route::get('/games/{game}', [GameController::class, 'show'])->name('games.show')
 Route::get('/games/category/{name}', [GameController::class, 'categoryFilter'])->name('games.category');
 Route::get('/filter', [GameController::class, 'filter'])->name('games.filter');
 
-Route::get('/games/create', [GameController::class, 'create'])->name('games.create');
-Route::post('/games', [GameController::class, 'store'])->name('games.store');
+Route::get('/admin/games', [GameController::class, 'gameList'])->name('games.admin.list');
+Route::get('/admin/games/create', [GameController::class, 'create'])->name('games.create');
+Route::post('/admin/games', [GameController::class, 'store'])->name('games.store');
 
-Route::get('/games/edit', [GameController::class, 'edit'])->name('games.edit');
-Route::put('/games/{game}', [GameController::class, 'update'])->name('games.update');
+Route::get('/admin/games/edit/{game}', [GameController::class, 'edit'])->name('games.edit');
+Route::put('/admin/games/{game}', [GameController::class, 'update'])->name('games.update');
 
-Route::delete('/games/{game}', [GameController::class, 'destroy'])->name('games.delete');
+Route::delete('/admin/games/{game}', [GameController::class, 'destroy'])->name('games.delete');
 
 Route::post('/games/comment', [GameController::class, 'comment'])->name('games.comment'); //$game->users()->attach(2,['is_purchased' => false, 'comment'=> 'dasdsa'])
 
@@ -41,21 +42,23 @@ Route::post('/email', [GameController::class, 'email'])->name('games.email');
 //categories
 Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
 
-Route::get('/categories/create', [CategoryController::class, 'create'])->name('categories.create');
-Route::post('/categories', [CategoryController::class, 'store'])->name('categories.store');
+Route::get('/admin/categories', [CategoryController::class, 'categoryList'])->name('categories.admin.list');
+Route::get('/admin/categories/create', [CategoryController::class, 'create'])->name('categories.create');
+Route::post('/admin/categories', [CategoryController::class, 'store'])->name('categories.store');
 
-Route::get('/categories/edit', [CategoryController::class, 'edit'])->name('categories.edit');
-Route::put('/categories/{category}', [CategoryController::class, 'update'])->name('categories.update');
+Route::get('/admin/categories/edit/{category}', [CategoryController::class, 'edit'])->name('categories.edit');
+Route::put('/admin/categories/{category}', [CategoryController::class, 'update'])->name('categories.update');
 
-Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->name('categories.delete');
+Route::delete('/admin/categories/{category}', [CategoryController::class, 'destroy'])->name('categories.delete');
 
 //Plataforms
 Route::get('/plataforms', [PlataformController::class, 'index'])->name('plataforms.index');
 
-Route::get('/plataforms/create', [PlataformController::class, 'create'])->name('plataforms.create');
-Route::post('/plataforms', [PlataformController::class, 'store'])->name('plataforms.store');
+Route::get('/admin/plataforms', [PlataformController::class, 'plataformList'])->name('plataforms.admin.list');
+Route::get('/admin/plataforms/create', [PlataformController::class, 'create'])->name('plataforms.create');
+Route::post('/admin/plataforms', [PlataformController::class, 'store'])->name('plataforms.store');
 
-Route::get('/plataforms/edit', [PlataformController::class, 'edit'])->name('plataforms.edit');
-Route::put('/plataforms/{category}', [PlataformController::class, 'update'])->name('plataforms.update');
+Route::get('/admin/plataforms/edit/{plataform}', [PlataformController::class, 'edit'])->name('plataforms.edit');
+Route::put('/admin/plataforms/{plataform}', [PlataformController::class, 'update'])->name('plataforms.update');
 
-Route::delete('/plataforms/{category}', [PlataformController::class, 'destroy'])->name('plataforms.delete');
+Route::delete('/admin/plataforms/{plataform}', [PlataformController::class, 'destroy'])->name('plataforms.delete');
